Automatic Generation of Tabs

It allows to automatically generate the code necessary to group multiple Rmarkdown chunks into tabs. Concatenating all the chunks into a string that can be later knitted and rendered.

automagic_tabs(input_data, panel_name, .output, .layout = NULL, ...)

Arguments

input_data
tibble with at least 2 columns, one for the title of the tabs and another with the output to be displayed.

panel_name
string with the name of the ID column.

.output
string with the name of the column of the output.

.layout
string that represents the layout of the tabs, can take the values "l-body", "l-body-outset", "l-page" and "l-screen". By default the value is NULL and takes "l-body" as parameter.

...
additional parameters that correspond to all those available in rmarkdown chunks (fig.align, fig.width, ...).

Value

concatenated string of all automatically generated chunks.

Details

given a tiblle, which must contain an "ID" column (representing the title of the tabs) and another column that stores the output to be generated (plot, text, code, ...), a string is automatically generated which can be later rendered in a Rmarkdown document.

See also

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(ggplot2)
library(purrr)
library(sknifedatar)

dataset <-
 tibble::tibble(ID = c("A","B"),
                numbers = list(rnorm(5), runif(5))) %>%
 mutate(plots = map(numbers,
                    ~ data.frame(plots = .x) %>% ggplot(aes(x=plots)) + geom_histogram()))

dataset

#> # A tibble: 2 x 3
#>   ID    numbers   plots 
#>   <chr> <list>    <list>
#> 1 A     <dbl [5]> <gg>  
#> 2 B     <dbl [5]> <gg>  

automagic_tabs(input_data = dataset, panel_name = "ID", .output = "plots")

#>   |                                                                              |                                                                      |   0%  |                                                                              |..............                                                        |  20%
#>    inline R code fragments
#> 
#> 
#> 
#> 
#> 
#> 

#> Quitting from lines 2-5 () 

#> Error in eval(parse_only(code), envir = envir): object 'dataset' not found

#> 

unlink("figure", recursive = TRUE)