Buenos Aires crimes data

Data set that records the date, time, type of crime and geolocation of crimes that occurred between 2017 and 2019. The data was extracted from the public repository of GCBA

crimes

Format

data frame with 100 rows y 9 columns:

id:

id

fecha:

date

franja_horaria:

hour from 0 to 23

tipo_delito:

crime type

subtipo_delito:

crime subtype

comuna:

commune

barrio:

neighborhood

lat:

latitude

long:

longitude

...

Source

http://data.buenosaires.gob.ar/dataset/delitos