Función para agregar valores NA a un data frame

Esta función permite agregar valores NA a un data frame, pudiendo seleccionar las columnas y la propòrcion de NAs deseados.

insert_na(data, columnas, p = 0.01, seed = 123)

Arguments

data
data frame

columnas
vector que indica el nombre de las columnas donde se agregaran los valores NA, ejemplo de formato c("X1","X2").

p
valor entre 0 y 1, indica la proporción de valores NA que se agregaran.

seed
semilla de números aleatorios.

Value

El data frame original suministrado a data, pero con los valores NA agregados en las columnas indicadas.

Examples

insert_na(data = iris, columnas = c("Sepal.Length","Petal.Length"), p = 0.25)

#> # A tibble: 150 x 5
#>    Sepal.Width Petal.Width Species Sepal.Length Petal.Length
#>          <dbl>       <dbl> <fct>          <dbl>        <dbl>
#>  1         3.5         0.2 setosa           5.1         NA  
#>  2         3           0.2 setosa          NA            1.4
#>  3         3.2         0.2 setosa           4.7          1.3
#>  4         3.1         0.2 setosa          NA            1.5
#>  5         3.6         0.2 setosa          NA            1.4
#>  6         3.9         0.4 setosa           5.4          1.7
#>  7         3.4         0.3 setosa           4.6          1.4
#>  8         3.4         0.2 setosa          NA            1.5
#>  9         2.9         0.2 setosa           4.4          1.4
#> 10         3.1         0.1 setosa           4.9          1.5
#> # … with 140 more rows