Fit Multiple Models to Multiple Time Series

This function allows multiple models to be adjusted over multiple time series, using models from the modeltime package.

modeltime_multifit(serie, .prop, ...)

Arguments

serie
nested time series.

.prop
series train/test partition ratio.

...
models or workflows to train (model_1, model2, ...).

Value

A list of 2 items. The first component is a tibble with a first column that contains the name of the series, and a second column called "nested_column" that stores the time series, then a column for each model where the trained models or workflows for each series are stored. The last 2 columns, "nested_model" and "calibration", store the "n" trained models for each series and the adjustment metrics on the test partition.

Details

The focus of this function is not related to panel series, it is oriented to multiple individual series. Receiving as the first argument "series" a set of nested series (for example through the nest function), then specifying a desired train/test partition ratio for series. The final input to the function are the models to be trained, simply by typing the name of the models separated by commas. The function admits as many models as required.

See also

Examples

# Generate "table_time" object
## libraries
library(modeltime)
library(rsample)
library(parsnip)
library(recipes)
library(workflows)
library(dplyr)
library(tidyr)
library(sknifedatar)

## Data
emae_series <- sknifedatar::emae_series
nested_serie <- emae_series %>% filter(date < '2008-02-01') %>% nest(nested_column=-sector)

## Recipes
recipe_1 = recipe(value ~ ., data = emae_series %>% select(-sector)) %>%
  step_date(date, features = c("month", "quarter", "year"), ordinal = TRUE)

## Models
m_naive_reg <- naive_reg() %>%  set_engine("naive")

m_nnetar <- workflow() %>%
  add_model(nnetar_reg() %>% set_engine("nnetar")) %>%
  add_recipe(recipe_1)

# modeltime_multifit
model_table_emae = modeltime_multifit(serie = nested_serie %>% head(2),
                                      .prop = 0.8,
                                      m_naive_reg ,
                                      m_nnetar)

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> Warning: A correlation computation is required, but `estimate` is constant and has 0 standard deviation, resulting in a divide by 0 error. `NA` will be returned.

#> Warning: A correlation computation is required, but `estimate` is constant and has 0 standard deviation, resulting in a divide by 0 error. `NA` will be returned.

model_table_emae

#> $table_time
#> # A tibble: 2 x 6
#>   sector   nested_column   m_naive_reg m_nnetar  nested_model    calibration    
#>   <chr>    <list>          <list>      <list>    <list>          <list>         
#> 1 Comercio <tibble [49 × … <fit[+]>    <workflo… <model_time [2… <model_time [2…
#> 2 Enseñan… <tibble [49 × … <fit[+]>    <workflo… <model_time [2… <model_time [2…
#> 
#> $models_accuracy
#> # A tibble: 4 x 10
#>   name_serie .model_id .model_desc    .type   mae  mape  mase smape  rmse    rsq
#>   <chr>          <int> <chr>          <chr> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 Comercio           1 NAIVE          Test   6.16  4.51  1.46  4.39  6.70 NA    
#> 2 Comercio           2 NNAR(1,1,10)[… Test   5.07  3.65  1.20  3.73  5.38  0.829
#> 3 Enseñanza          1 NAIVE          Test   4.46  3.75  1.96  3.84  5.00 NA    
#> 4 Enseñanza          2 NNAR(1,1,10)[… Test   3.23  2.73  1.42  2.76  3.45  0.377
#>