Función para reajustar el o los modelos para múltiples series de tiempo

Esta función permite aplicar la función "modeltime_refit()" de modeltime a múltiples series y modelos.

modeltime_multirefit(models_table)

Arguments

models_table
tibble que proviene del objeto "table_time" de la salida de la función "modeltime_multifit".

Value

Devuelve el objeto "table_time" recalibrado.

Details

Toma como input el objeto "table_time" de la salida de la función "modeltime_multifit" y devuelve el mismo objeto pero entrenado en todo el período.

Examples

library(sknifedatar)
library(modeltime)

data("table_time")

table_time_refit <- modeltime_multirefit(models_table = table_time)

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

table_time_refit

#> # A tibble: 3 x 7
#>   sector       nested_column   m_auto_arima m_stlm_arima m_nnetar nested_model  
#>   <chr>        <list>          <list>       <list>       <list>   <list>        
#> 1 Comercio     <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> 2 Enseñanza    <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> 3 Administrac… <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> # … with 1 more variable: calibration <list>