Evaluación de múltiples métricas y predicciones

Para un conjunto de predicciones de distintos modelos, permite evaluar múltiples métricas y devolver los resultados en un formato tabular que facilita la comparación de las predicciones.

multieval(data, observed, predictions, metrica, plot_view = FALSE)

Arguments

data
data frame con las predicciones, debe tener como mínimo la columna con los datos observados y al menos una columna que haga referencia a las predicciones de un modelo.

observed
string con el nombre de la columna que contiene los datos observados.

predictions
string o vector de strings las columnas donde se almacenan las predicicones.

metrica
métrica o conjunto de métricas que se desean evaluar, las métricas hacen referencias a las permitidas por el paquete yardstick de tidymodels.

plot_view
TRUE para hacer un gráfico de las métricas resultantes.

Value

data frame con 4 columnas, siendo las mismas las métricas de evaluación, el estimador usado, el valor de la métrica y el nombre del modelo.

See also

Examples

set.seed(123)
predictions =
  data.frame(truth = runif(100),
             predict_model_1 = rnorm(100, mean = 1,sd =2),
             predict_model_2 = rnorm(100, mean = 0,sd =2))

multieval(data = predictions,
          observed = "truth",
          predictions = c("predict_model_1","predict_model_2"),
          metrica = list(rmse = yardstick::rmse,
                         rsq = yardstick::rsq,
                         mae = yardstick::mae))

# Output ----------------------
#  A tibble: 6 x 4
#  .metric .estimator .estimate model
#  <chr>   <chr>          <dbl> <chr>
#1 mae     standard    1.59     predict_model1
#2 mae     standard    1.61     predict_model2
#3 rmse    standard    1.99     predict_model1
#4 rmse    standard    1.95     predict_model2
#5 rsq     standard    0.000704 predict_model1
#6 rsq     standard    0.00115  predict_model2

# Alternativa utilizando "erer::listn" en lugar de
# "list" para pasar las méticas como argumentos
#
# install.packages("erer")
#
# multieval(data = predictions,
#          observed = "truth",
#          predictions = c("predict_model1","predict_model2"),
#          metrica = erer::listn(rmse, rsq, mae))