Puntos geolocalizados dentro de un radio de metros de cercanía

Dado dos conjuntos de puntos geolocalizados, esta función permite determinar para cada punto del primer conjunto de datos, cuál o cuáles de los puntos del segundo conjunto de datos están dentro de un radio de metros determinado.

pertenencia_punto(data, referencia, metros)

Arguments

data
data frame que representa los puntos del primer conjunto de datos,

referencia
data frame que representa los segundos puntos de datos, sobre los cuales se evaluará cuáles de ellos están dentro del radio de cercanía de los puntos del primer conjunto de datos, las columnas de geolocalización deben estar nombradas como "long" y "lat".

metros
radio en metros para evaluar la cercanía.

Value

Una lista del mismo tamaño que el primer conjunto de datos, cada elemento tiene como valor el o los puntos del segundo conjunto de datos que estén dentro del radio de cercanía.

Details

Las columnas de geolocalización de ambos conjuntos de datos deben estar nombradas como "long" y "lat".

See also

Examples

pertenencia_punto(data = crimes, referencia = intercepcion_calles[1:10,], metros = 150)

#> $`1`
#> [1] 0
#> 
#> $`2`
#> [1] 0
#> 
#> $`3`
#> [1] 0
#> 
#> $`4`
#> [1] 0
#> 
#> $`5`
#> [1] 0
#> 
#> $`6`
#> [1] 0
#> 
#> $`7`
#> [1] 0
#> 
#> $`8`
#> [1] 0
#> 
#> $`9`
#> [1] 0
#> 
#> $`10`
#> [1] 0
#> 
#> $`11`
#> [1] 0
#> 
#> $`12`
#> [1] 0
#> 
#> $`13`
#> [1] 0
#> 
#> $`14`
#> [1] 0
#> 
#> $`15`
#> [1] 0
#> 
#> $`16`
#> [1] 0
#> 
#> $`17`
#> [1] 0
#> 
#> $`18`
#> [1] 0
#> 
#> $`19`
#> [1] 0
#> 
#> $`20`
#> [1] 0
#> 
#> $`21`
#> [1] 0
#> 
#> $`22`
#> [1] 0
#> 
#> $`23`
#> [1] 0
#> 
#> $`24`
#> [1] 0
#> 
#> $`25`
#> [1] 0
#> 
#> $`26`
#> [1] 0
#> 
#> $`27`
#> [1] 0
#> 
#> $`28`
#> [1] 0
#> 
#> $`29`
#> [1] 0
#> 
#> $`30`
#> [1] 0
#> 
#> $`31`
#> [1] 0
#> 
#> $`32`
#> [1] 0
#> 
#> $`33`
#> [1] 0
#> 
#> $`34`
#> [1] 0
#> 
#> $`35`
#> [1] 0
#> 
#> $`36`
#> [1] 0
#> 
#> $`37`
#> [1] 0
#> 
#> $`38`
#> [1] 0
#> 
#> $`39`
#> [1] 0
#> 
#> $`40`
#> [1] 0
#> 
#> $`41`
#> [1] 0
#> 
#> $`42`
#> [1] 0
#> 
#> $`43`
#> [1] 0
#> 
#> $`44`
#> [1] 0
#> 
#> $`45`
#> [1] 0
#> 
#> $`46`
#> [1] 0
#> 
#> $`47`
#> [1] 0
#> 
#> $`48`
#> [1] 0
#> 
#> $`49`
#> [1] 0
#> 
#> $`50`
#> [1] 0
#> 
#> $`51`
#> [1] 0
#> 
#> $`52`
#> [1] 0
#> 
#> $`53`
#> [1] 0
#> 
#> $`54`
#> [1] 0
#> 
#> $`55`
#> [1] 0
#> 
#> $`56`
#> [1] 0
#> 
#> $`57`
#> [1] 0
#> 
#> $`58`
#> [1] 0
#> 
#> $`59`
#> [1] 0
#> 
#> $`60`
#> [1] 0
#> 
#> $`61`
#> [1] 0
#> 
#> $`62`
#> [1] 0
#> 
#> $`63`
#> [1] 0
#> 
#> $`64`
#> [1] 0
#> 
#> $`65`
#> [1] 0
#> 
#> $`66`
#> [1] 0
#> 
#> $`67`
#> [1] 0
#> 
#> $`68`
#> [1] 0
#> 
#> $`69`
#> [1] 0
#> 
#> $`70`
#> [1] 0
#> 
#> $`71`
#> [1] 0
#> 
#> $`72`
#> [1] 0
#> 
#> $`73`
#> [1] 0
#> 
#> $`74`
#> [1] 0
#> 
#> $`75`
#> [1] 0
#> 
#> $`76`
#> [1] 0
#> 
#> $`77`
#> [1] 0
#> 
#> $`78`
#> [1] 0
#> 
#> $`79`
#> [1] 0
#> 
#> $`80`
#> [1] 0
#> 
#> $`81`
#> [1] 0
#> 
#> $`82`
#> [1] 0
#> 
#> $`83`
#> [1] 0
#> 
#> $`84`
#> [1] 0
#> 
#> $`85`
#> [1] 0
#> 
#> $`86`
#> [1] 0
#> 
#> $`87`
#> [1] 0
#> 
#> $`88`
#> [1] 0
#> 
#> $`89`
#> [1] 0
#> 
#> $`90`
#> [1] 0
#> 
#> $`91`
#> [1] 0
#> 
#> $`92`
#> [1] 0
#> 
#> $`93`
#> [1] 0
#> 
#> $`94`
#> [1] 0
#> 
#> $`95`
#> [1] 0
#> 
#> $`96`
#> [1] 0
#> 
#> $`97`
#> [1] 0
#> 
#> $`98`
#> [1] 0
#> 
#> $`99`
#> [1] 0
#> 
#> $`100`
#> [1] 0
#>