Función para obtener el mejor modelo a partir de una modeltime table

Esta función permite seleccionar el mejor modelo para cada serie, en función de determinada métrica de evaluación.

modeltime_multibestmodel(.table, .metric = "mase", .optimization = which.min)

Arguments

.table
tibble que proviene del objeto "table_time" de la salida de la función "modeltime_multifit".

.metric
métrica de evaluación, proviene de "modeltime_accuracy": 'mae', 'mape','mase','smape','rmse','rsq'.

.optimization
Es posible minimizar una métrica de error ('which.min') o maximizar el rsq ('which.max').

Value

tibble table_time filtrada por el mejor modelo.

Details

Esta función toma el objeto "table_time" de la salida de la función "modeltime_multifit", posteriomente se selecciona el mejor modelo en función de la métrica seleccionada.

Examples

library(sknifedatar)
data("table_time")

best_model_emae <- modeltime_multibestmodel(.table=table_time,
                                            .metric=rmse,
                                            .optimization = which.min)

#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 

best_model_emae

#> # A tibble: 3 x 8
#>   sector       nested_column   m_auto_arima m_stlm_arima m_nnetar nested_model  
#>   <chr>        <list>          <list>       <list>       <list>   <list>        
#> 1 Comercio     <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> 2 Enseñanza    <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> 3 Administrac… <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> # … with 2 more variables: calibration <list>, best_model <list>