Función para obtener el mejor modelo a partir de una modeltime table

Esta función permite seleccionar el mejor modelo para cada serie, en función de determinada métrica de evaluación.

modeltime_multibestmodel(.table, .metric = "mase", .optimization = which.min)

Arguments

.table
tibble que proviene del objeto "table_time" de la salida de la función "modeltime_multifit".

.metric
métrica de evaluación, proviene de "modeltime_accuracy": 'mae', 'mape','mase','smape','rmse','rsq'.

.optimization
Es posible minimizar una métrica de error ('which.min') o maximizar el rsq ('which.max').

Value

tibble table_time filtrada por el mejor modelo.

Details

Esta función toma el objeto "table_time" de la salida de la función "modeltime_multifit", posteriomente se selecciona el mejor modelo en función de la métrica seleccionada.

Examples

# Generate "table_time" object
## libraries
library(modeltime)
library(rsample)

#> 
#> Attaching package: ‘rsample’

#> The following object is masked from ‘package:sknifedatar’:
#> 
#>     sliding_window

library(parsnip)
library(recipes)

#> 
#> Attaching package: ‘recipes’

#> The following object is masked from ‘package:stats’:
#> 
#>     step

library(workflows)
library(dplyr)
library(tidyr)
library(sknifedatar)

## Data
emae_series <- sknifedatar::emae_series
nested_serie <- emae_series %>% filter(date < '2008-02-01') %>% nest(nested_column=-sector)

## Recipes
recipe_1 = recipe(value ~ ., data = emae_series %>% select(-sector)) %>%
  step_date(date, features = c("month", "quarter", "year"), ordinal = TRUE)

## Models
m_ets <- workflow() %>%
  add_model(exp_smoothing() %>% set_engine('ets')) %>%
  add_recipe(recipe_1)

m_nnetar <- workflow() %>%
  add_model(nnetar_reg() %>% set_engine("nnetar")) %>%
  add_recipe(recipe_1)

# modeltime_multifit
model_table_emae = modeltime_multifit(serie = nested_serie %>% head(2),
                                      .prop = 0.8,
                                      m_ets,
                                      m_nnetar)

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

table_time <- model_table_emae$table_time

# best_model_emae
best_model_emae <- modeltime_multibestmodel(.table=table_time,
                                            .metric=rmse,
                                            .optimization = which.min)

best_model_emae

#> # A tibble: 2 x 7
#>   sector  nested_column  m_ets   m_nnetar nested_model  calibration   best_model
#>   <chr>   <list>         <list>  <list>   <list>        <list>        <list>    
#> 1 Comerc… <tibble [49 ×… <workf… <workfl… <model_time … <model_time … <int [1]> 
#> 2 Enseña… <tibble [49 ×… <workf… <workfl… <model_time … <model_time … <int [1]>