Ajuste de múltiples modelos a múltiples series de tiempo

Esta función permite ajustar múltiples modelos sobre múltiples series de tiempo, utilizando los modelos del paquete modeltime.

modeltime_multifit(serie, .prop, ...)

Arguments

serie
series de tiempo anidadas.

.prop
proporción de partición de las series en train/test.

...
modelos o workflows a entrenar (model_1,model2,...).

Value

una lista con 2 elementos. El primer componente es un tibble que contiene una primera columna que contiene el nombre de las series, seguidamente la comuna "nested_column" que almacena las series de tiempo, luego una columna para cada modelo suministrado donde se almacenan los modelos o workflows entrenados para cada serie.

    Finalmente las columnas "nested_model" y "calibration" que guardan los "n" modelos entrenados para cada serie y las métricas de ajuste sobre la
    partición de test.

Details

El enfoque de esta función no está relacionado con series de panel, está orientada a múltiples series individuales. Recibiendo como primer argumento "serie" un conjunto de series anidadas (por ejemplo a través de la función nest), luego especificando una proporción deseada para particionar cada una de las series en train/test.

     Por último se deben suministrar los modelos a entrenar, simplemente escribiendo el nombre de los modelos separados
     por comas, la función admite tantos modelos como se requieran.

See also

Examples

# libraries
library(modeltime)
library(rsample)

#> 
#> Attaching package: ‘rsample’

#> The following object is masked from ‘package:sknifedatar’:
#> 
#>     sliding_window

library(parsnip)
library(recipes)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> 
#> Attaching package: ‘recipes’

#> The following object is masked from ‘package:stats’:
#> 
#>     step

library(workflows)
library(dplyr)
library(tidyr)
library(sknifedatar)

# Data
data("emae_series")
nested_serie = emae_series %>% filter(date < '2020-02-01') %>% nest(nested_column=-sector)

# Recipes
recipe_1 = recipe(value ~ ., data = emae_series %>% select(-sector)) %>%
step_date(date, features = c("month", "quarter", "year"), ordinal = TRUE)

# Models
m_auto_arima <- arima_reg() %>% set_engine('auto_arima')

m_stlm_arima <- seasonal_reg() %>%
  set_engine("stlm_arima")

m_nnetar <- workflow() %>%
  add_recipe(recipe_1) %>%
  add_model(nnetar_reg() %>% set_engine("nnetar"))

# modeltime_multifit
model_table_emae = modeltime_multifit(serie = nested_serie %>% head(3),
                                      .prop = 0.8,
                                      m_auto_arima,
                                      m_stlm_arima,
                                      m_nnetar)

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

#> frequency = 12 observations per 1 year

model_table_emae

#> $table_time
#> # A tibble: 3 x 7
#>   sector       nested_column   m_auto_arima m_stlm_arima m_nnetar nested_model  
#>   <chr>        <list>          <list>       <list>       <list>   <list>        
#> 1 Comercio     <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> 2 Enseñanza    <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> 3 Administrac… <tibble [193 ×… <fit[+]>     <fit[+]>     <workfl… <model_time […
#> # … with 1 more variable: calibration <list>
#> 
#> $models_accuracy
#> # A tibble: 9 x 10
#>   name_serie  .model_id .model_desc .type   mae  mape   mase smape  rmse     rsq
#>   <chr>           <int> <chr>       <chr> <dbl> <dbl>  <dbl> <dbl> <dbl>   <dbl>
#> 1 Comercio            1 ARIMA(0,1,… Test  11.9   8.70  1.10   8.41 15.0  1.53e-4
#> 2 Comercio            2 SEASONAL D… Test  12.5   9.24  1.16   8.75 15.7  1.01e-4
#> 3 Comercio            3 NNAR(1,1,1… Test   8.27  5.97  0.768  5.83  9.01 4.37e-1
#> 4 Enseñanza           1 ARIMA(1,1,… Test   3.65  2.27  2.80   2.26  4.27 8.08e-2
#> 5 Enseñanza           2 SEASONAL D… Test   3.67  2.28  2.82   2.28  4.29 7.80e-2
#> 6 Enseñanza           3 NNAR(1,1,1… Test   4.80  2.97  3.69   2.92  5.18 4.76e-1
#> 7 Administra…         1 ARIMA(0,1,… Test   3.23  2.10  6.28   2.07  3.59 3.43e-1
#> 8 Administra…         2 SEASONAL D… Test   3.09  2.01  6.00   1.98  3.56 2.84e-1
#> 9 Administra…         3 NNAR(1,1,1… Test  10.2   6.60 19.8    6.38 10.3  1.10e-2
#>