Trasnformación de ventana deslizante movil

Esta función permite aplicar una trasnformación de ventana deslizante movil mensual sobre un conjunto de datos.

sliding_window(data, inicio, pliegues, variables)

Arguments

data
data frame que contiene conteos históricos de distintos eventos en espacios de tiempo mensuales. Cada fila es una observación única y las columnas corresponden a los distintos meses de estudio. Las variables deben tener palabras claves para poder seleccionarlas en conjunto. Para ver un ejemplo de la estructura de los datos, puede consultarse el dataset tal contenido en este paquete.

inicio
mes inicial, formato numerico entero

pliegues
vector que inicia en 1 y termina en el número de periodos que se desee recorrer.

variables
una palabra o vector que permita seleccionar las variables en conjunto e implementar la función para cada grupo

Value

Un data frame con el ID de la observaciones y los distintos espacios de tiempo de conteo calculados por variables

Details

El funcionamiento es el siguiente, es seleccionado el mes intermedio “t” de todo el periodo de estudio, posteriormente se calcula el número de eventos ocurridos para cada obervacion en el mes anterior, en los últimos 3 meses, 6 meses, 12 meses y el mismo mes del año anterior.

El procedimiento descrito anteriormente es replicado de manera móvil, es decir, rodando la ventana temporal desde t+1 hasta n, siendo n el último mes de estudio. Para ver un caso de uso real, visitar Crime analysis with tidymodels

See also

Examples

pliegues = 1:13
names(pliegues) = pliegues

variables = c("delitos", "temperatura", "mm_agua", "lluvia", "viento")
names(variables) = variables

sliding_window(data = data_crime_clime %>% dplyr::select(-c(long,lat)),
               inicio = 13,
               pliegues = pliegues,
               variables = variables)

#> # A tibble: 26,299 x 32
#>    id    pliegue delitos_last_ye… delitos_last_12 delitos_last_6 delitos_last_3
#>    <chr>   <int>            <dbl>           <dbl>          <dbl>          <dbl>
#>  1 esqu…       1                1              73             41             20
#>  2 esqu…       1                5             106             47             19
#>  3 esqu…       1                4              25             11              4
#>  4 esqu…       1                0               4              1              1
#>  5 esqu…       1                0              31             16              7
#>  6 esqu…       1                1              11              6              3
#>  7 esqu…       1                1              16             11              4
#>  8 esqu…       1                1              19              8              5
#>  9 esqu…       1                0               9              6              4
#> 10 esqu…       1                3              27             14             10
#> # … with 26,289 more rows, and 26 more variables: delitos_last_1 <dbl>,
#> #   delitos <dbl>, temperatura_last_year <dbl>, temperatura_last_12 <dbl>,
#> #   temperatura_last_6 <dbl>, temperatura_last_3 <dbl>,
#> #   temperatura_last_1 <dbl>, temperatura <dbl>, mm_last_year <dbl>,
#> #   mm_last_12 <dbl>, mm_last_6 <dbl>, mm_last_3 <dbl>, mm_last_1 <dbl>,
#> #   mm <dbl>, dias_last_year <dbl>, dias_last_12 <dbl>, dias_last_6 <dbl>,
#> #   dias_last_3 <dbl>, dias_last_1 <dbl>, dias <dbl>, veloc_last_year <dbl>,
#> #   veloc_last_12 <dbl>, veloc_last_6 <dbl>, veloc_last_3 <dbl>,
#> #   veloc_last_1 <dbl>, veloc <dbl>